change to long long int for sigset_t - might be missing for C++, but not with compiler we're using?

+++ src/gnuwin32/fixed/h/psignal.h
@@ -72,12 +72,12 @@
 */
 typedef	void (*sighandler_t)(int nSig);
 
-/* Define sigset_t as long int, regardless of what mingw wants, which
+/* Define sigset_t as long long int, regardless of what mingw wants, which
    is confused with whether or not _POSIX is defined.  Note that uses
    of this type are here in psignal.c, so we should be able to decide
    for ourselves what it is. */
 
-typedef long int sigset_t;
+typedef long long int sigset_t;
 
 
 /*
